m15/pyproject.toml
[tool.poetry]
name = "m15"
version = "0.1.0"
description = ""
authors = ["Yuriy Kuchma <krabatua@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.1"
uvicorn = {extras = ["standard"], version = "^0.19.0"}
SQLAlchemy = "^1.4.42"
psycopg2 = "^2.9.4"
alembic = "^1.8.1"
bcrypt = "^4.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
